all basic funcitons done
